+++ crates/nostr-sdk/src/client/mod.rs
@@ -87,7 +87,6 @@ impl Client {
             builder.signer,
             builder.admit_policy,
             builder.opts.nip42_auto_authentication,
-            builder.opts.min_pow_difficulty,
         );
 
         // Construct client
@@ -106,10 +105,11 @@ impl Client {
     /// Update minimum POW difficulty for received events
     ///
     /// Events with a POW lower than the current value will be ignored to prevent resources exhaustion.
-    #[inline]
-    pub fn update_min_pow_difficulty(&self, difficulty: u8) {
-        self.state().set_pow(difficulty);
-    }
+    #[deprecated(
+        since = "0.40.0",
+        note = "This no longer works, please use `AdmitPolicy` instead."
+    )]
+    pub fn update_min_pow_difficulty(&self, _difficulty: u8) {}
 
     /// Auto authenticate to relays (default: true)
     ///