+++ crates/nostr-sdk/src/client/options.rs
@@ -16,7 +16,6 @@ use nostr_relay_pool::prelude::*;
 #[derive(Debug, Clone)]
 pub struct Options {
     pub(super) autoconnect: bool,
-    pub(super) min_pow_difficulty: u8,
     pub(super) nip42_auto_authentication: bool,
     pub(super) gossip: bool,
     #[cfg(not(target_arch = "wasm32"))]
@@ -30,7 +29,6 @@ impl Default for Options {
     fn default() -> Self {
         Self {
             autoconnect: false,
-            min_pow_difficulty: 0,
             nip42_auto_authentication: true,
             gossip: false,
             #[cfg(not(target_arch = "wasm32"))]
@@ -59,9 +57,11 @@ impl Options {
     }
 
     /// Minimum POW difficulty for received events (default: 0)
-    #[inline]
-    pub fn min_pow(mut self, difficulty: u8) -> Self {
-        self.min_pow_difficulty = difficulty;
+    #[deprecated(
+        since = "0.40.0",
+        note = "This no longer works, please use `AdmitPolicy` instead."
+    )]
+    pub fn min_pow(self, _difficulty: u8) -> Self {
         self
     }